Use pkgutil '--pkg-info' instead of '--pkgs'.
"pkgutil --pkgs='PACKAGE_ID'" treats PACKAGE_ID like a regular
expression. The dmg cookbook documentation implies that the value of
PACKAGE_ID is a literal package id, not a regular expression. In this
case, it makes more sense to use "pkgutil --pkg-info 'PACKAGE_ID'",
which matches on the literal package id.